v0.3.1.4 fix schema data types

+++ backend/db/schema/Users.js
@@ -1,7 +1,7 @@
 import {
   mysqlTable,
   varchar,
-  bigint,
+  int,
   boolean,
   timestamp,
 } from "drizzle-orm/mysql-core";
